Reduced/removed usage of reflection and traverse
Considering Publicizer is one of the NuGet packages, it made sense to me to replace calls to vanilla methods (patch_DamageWorker_AddInjury) and access to fields (Patch_PawnGenerator_GenerateGearFor).

I've also included DualWield to be publicized as well, and invoked its method directly as well (patch_DrawEquipment). If you don't want to publicize it, I can instead replace it with MethodInvoker/FastInvokeHandler from Harmony.

On top of that, I removed redundant call to FieldInfo.GetValue (yayoCombat). This won't have much impact as it's a one-time thing, but why not include it as well.

There's one more usage of reflection that I've left alone, as it seems to be one time call anyway (PawnRenderer_override). However, I've moved the second null check into the first one to avoid redundant double call for the null check there since after the first one returns false, the second one is guaranteed to as well.